+++ docs/changes.rst
@@ -28,6 +28,7 @@ Unreleased.
   identifying layer visibility.
 - Added :attr:`Image.profiles <wand.image.ProfileDict>` dictionary attribute. [:issue:`249`]
 - Moved :mod:`collections.abc` to :attr:`wand.compat.abc` for Python-3.8. [:issue:`394` by Tero Vuotila]
+- Update :mod:`wand.display` to use Python3 compatible :func:`print()` function. [:issue:`395` by Tero Vuotila]
 
 
 .. _changelog-0.5.0: